examples/deep-research/pyproject.toml
[project]
name = "deep_research"
version = "0.1.0"
description = "Deep Research agent using LangCrew"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "langcrew>=0.1.1",
    "langchain-community>=0.3.0",
    "google-search-results>=2.4.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
deep_research = "deep_research.main:run"

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

